requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
diff-match-patch==20121119
Django==2.0.3
django-admin-interface==0.6.3
-e git+https://github.com/thienkimlove/django-admin-menu.git@35e2945849fbe59903403c20a19b3f479ffb6ec4#egg=django_admin_menu
django-appconf==1.0.1
-e git+https://github.com/yourlabs/django-autocomplete-light.git@037a83a3c21c425fa271775f56b7d22abd30d373#egg=django_autocomplete_light
django-bootstrap4==0.0.6
django-ckeditor==5.4.0
django-classy-tags==0.6.2
django-colorfield==0.1.14
django-constance==2.2.0
django-countries==5.2
django-datatables-view==1.16.0
django-debug-toolbar==1.9.1
django-djconfig==0.6.0
django-etc==0.10.0
django-extensions==2.0.6
django-filebrowser-no-grappelli==3.7.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-gm2m==0.6.1
django-import-export==1.0.0
django-js-asset==1.0.0
django-jsoneditor==0.0.12
django-logentry-admin==1.0.3
django-picklefield==1.0.0
django-select2==6.0.2
django-siteprefs==0.7.0
django-smart-selects==1.5.4
django-taggit==0.22.2
django-taggit-autosuggest==0.3.3
django-widget-tweaks==1.4.2
docutils==0.14
et-xmlfile==1.0.1
fake-factory==9999.9.9
idna==2.6
jdcal==1.3
jsonfield==2.0.2
libsass==0.14.5
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.0
packaging==17.1
pafy==0.3.82
persisting-theory==0.2.1
Pillow==5.0.0
#pkg-resources==0.0.0
pycountry==18.5.26
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
uWSGI==2.0.17
xlrd==1.1.0
xlwt==1.3.0
